Move library/include checks and configure.h to toplevel
Do all the checks in toplevel CMakeLists.txt and create a configure.h
that can be used by both the client and the server code. TODO: The
remaining checks in src/CMakeLists.txt could be moved to toplevel as
well. Also using md5c.c if neither OpenSSL nor GnuTLS were found is not
working yet.