Fix crashing system name display on unknown Windows version
The calling GKrellM code always expects a system name that can be split
at the first space character to provide kernel name and kernel version.
Return a string containing a space to avoid a crash on any Windows
version the code does not know about (or in case querying the version
failed).